requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.14.0
apache-airflow-providers-http==4.9.0
apache-airflow-providers-postgres==5.10.0

# Testing
pytest==7.4.0
pytest-mock==3.12.0

# Utils
requests==2.31.0
boto3==1.34.0